+++ src/components/common/Header/AllMembersHeader.tsx
@@ -1,0 +1,125 @@
+import { useState, ChangeEvent } from "react";
+import {
+  Button,
+  FormControl,
+  InputLabel,
+  MenuItem,
+  Select,
+  SelectChangeEvent,
+  Stack,
+  styled,
+  TextField,
+} from "@mui/material";
+import { toast } from "react-toastify";
+import { allMemberApi } from "@/apis/allMemberApi";
+import { memberInfoSelectMenu } from "@/constants/table";
+import {
+  useAllMembersSearchInfoDispatch,
+  useAllMembersSearchInfoState,
+} from "@/hooks/contexts/useAllMembersSearchInfoContext";
+import { downloadExcelFile } from "@/utils/excel";
+
+export default function AllMembersHeader() {
+  const [selectedMemberInfoVariant, setSelectedMemberInfoVariant] = useState<number>(1);
+
+  const { setSearchText, setSearchVariant, setPaginationModel } = useAllMembersSearchInfoDispatch();
+  const { searchText } = useAllMembersSearchInfoState();
+
+  const handleClickExcelDownloadButton = async () => {
+    try {
+      const data = await allMemberApi.getMemberInfoExcel();
+      downloadExcelFile(data, "members");
+    } catch (error) {
+      toast.error("오류가 발생했습니다.");
+    }
+  };
+
+  const handleResetPage = () => {
+    setPaginationModel(prevPaginationModel => ({
+      ...prevPaginationModel,
+      page: 0,
+    }));
+  };
+
+  const handleChangeSelectMemberInfoVariant = (e: SelectChangeEvent<unknown>) => {
+    const targetIndex = (e.target.value as number) - 1;
+
+    setSearchVariant?.(memberInfoSelectMenu[targetIndex]["type"]);
+    setSearchText?.("");
+    setSelectedMemberInfoVariant(targetIndex + 1);
+    handleResetPage();
+  };
+
+  const handleChangeText = (e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
+    const text = e.target.value;
+    text.length === 1 && handleResetPage();
+
+    setSearchText?.(text);
+  };
+
+  return (
+    <StyledHeaderWrapper>
+      <StyledHeaderLeftColWrapper>
+        <StyledFormWrapper>
+          <InputLabel>Type</InputLabel>
+          <Select
+            label="Type"
+            value={selectedMemberInfoVariant}
+            onChange={handleChangeSelectMemberInfoVariant}
+          >
+            {memberInfoSelectMenu.map(title => (
+              <MenuItem value={title.value} key={title.value}>
+                {title.name}
+              </MenuItem>
+            ))}
+          </Select>
+        </StyledFormWrapper>
+        <StyledTextField
+          label="search"
+          variant="outlined"
+          placeholder="name, email, etc.."
+          value={searchText}
+          onChange={handleChangeText}
+        />
+      </StyledHeaderLeftColWrapper>
+      <StyledHeaderRightColWrapper>
+        <StyledExcelDownloadButton variant="outlined" onClick={handleClickExcelDownloadButton}>
+          엑셀 다운로드
+        </StyledExcelDownloadButton>
+      </StyledHeaderRightColWrapper>
+    </StyledHeaderWrapper>
+  );
+}
+
+const StyledHeaderWrapper = styled(Stack)({
+  gap: 20,
+  flexDirection: "row",
+  justifyContent: "space-between",
+  alignItems: "center",
+  flexWrap: "wrap",
+  padding: "16px",
+});
+
+const StyledHeaderLeftColWrapper = styled(Stack)({
+  gap: 10,
+  flexDirection: "row",
+  alignItems: "center",
+  flexWrap: "wrap",
+});
+
+const StyledHeaderRightColWrapper = styled(Stack)({
+  justifyContent: "space-between",
+});
+
+const StyledFormWrapper = styled(FormControl)({
+  width: "180px",
+});
+
+const StyledTextField = styled(TextField)({
+  width: 200,
+});
+
+const StyledExcelDownloadButton = styled(Button)({
+  padding: "8px 22px",
+  height: "32px",
+});
