+++ src/components/common/Header/PendingMembersHeader.tsx
@@ -1,0 +1,128 @@
+import { useState, ChangeEvent } from "react";
+import {
+  FormControl,
+  InputLabel,
+  MenuItem,
+  Select,
+  SelectChangeEvent,
+  Stack,
+  styled,
+  TextField,
+} from "@mui/material";
+import { memberInfoSelectMenu, memberTypeSelectMenu } from "@/constants/table";
+import {
+  usePendingMembersSearchInfoDispatch,
+  usePendingMembersSearchInfoState,
+} from "@/hooks/contexts/usePendingMembersSearchInfoContext";
+
+export default function PendingMembersHeader() {
+  const [selectedMemberInfoVariant, setSelectedMemberInfoVariant] = useState<number>(1);
+  const [selectedMemberVariant, setSelectedMemberVariant] = useState<number>(1);
+
+  const { setSearchText, setSearchVariant, setPaginationModel, setMemberVariant } =
+    usePendingMembersSearchInfoDispatch();
+  const { searchText } = usePendingMembersSearchInfoState();
+
+  const handleResetPage = () => {
+    setPaginationModel(prevPaginationModel => ({
+      ...prevPaginationModel,
+      page: 0,
+    }));
+  };
+
+  const handleChangeSelectMemberInfoVariant = (e: SelectChangeEvent<unknown>) => {
+    const targetIndex = (e.target.value as number) - 1;
+
+    setSearchVariant?.(memberInfoSelectMenu[targetIndex]["type"]);
+    setSearchText?.("");
+    setSelectedMemberInfoVariant(targetIndex + 1);
+    handleResetPage();
+  };
+
+  const handleChangeSelectMemberVariant = (e: SelectChangeEvent<unknown>) => {
+    const targetIndex = (e.target.value as number) - 1;
+
+    setMemberVariant?.(memberTypeSelectMenu[targetIndex]["type"]);
+    setSelectedMemberVariant(targetIndex + 1);
+    handleResetPage();
+  };
+
+  const handleChangeText = (e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
+    const text = e.target.value;
+    text.length === 1 && handleResetPage();
+
+    setSearchText?.(text);
+  };
+
+  return (
+    <StyledHeaderWrapper>
+      <StyledHeaderLeftColWrapper>
+        <StyledFormWrapper>
+          <InputLabel>Type</InputLabel>
+          <Select
+            label="Type"
+            value={selectedMemberInfoVariant}
+            onChange={handleChangeSelectMemberInfoVariant}
+          >
+            {memberInfoSelectMenu.map(title => (
+              <MenuItem value={title.value} key={title.value}>
+                {title.name}
+              </MenuItem>
+            ))}
+          </Select>
+        </StyledFormWrapper>
+        <StyledTextField
+          label="search"
+          variant="outlined"
+          placeholder="name, email, etc.."
+          value={searchText}
+          onChange={handleChangeText}
+        />
+      </StyledHeaderLeftColWrapper>
+      <StyledHeaderRightColWrapper>
+        <StyledFormWrapper>
+          <InputLabel>Type</InputLabel>
+          <Select
+            label="Type"
+            value={selectedMemberVariant}
+            onChange={handleChangeSelectMemberVariant}
+          >
+            {memberTypeSelectMenu.map(title => (
+              <MenuItem value={title.value} key={title.value}>
+                {title.name}
+              </MenuItem>
+            ))}
+          </Select>
+        </StyledFormWrapper>
+      </StyledHeaderRightColWrapper>
+    </StyledHeaderWrapper>
+  );
+}
+
+const StyledHeaderWrapper = styled(Stack)({
+  gap: 20,
+  flexDirection: "row",
+  justifyContent: "space-between",
+  alignItems: "center",
+  flexWrap: "wrap",
+  padding: "16px",
+});
+
+const StyledHeaderLeftColWrapper = styled(Stack)({
+  gap: 10,
+  flexDirection: "row",
+  alignItems: "center",
+  flexWrap: "wrap",
+});
+
+const StyledHeaderRightColWrapper = styled(Stack)({
+  justifyContent: "space-between",
+});
+
+const StyledFormWrapper = styled(FormControl)({
+  width: "180px",
+});
+
+const StyledTextField = styled(TextField)({
+  width: 200,
+});
