+++ src/pages/AllMembersPage.tsx
@@ -1,5 +1,5 @@
-import Header from "@/components/common/Header";
-import AllMemberInfoTable from "@/components/common/Table/AllMemberInfoTable";
+import AllMembersHeader from "@/components/common/Header/AllMembersHeader";
+import AllMembersInfoTable from "@/components/common/Table/AllMembersInfoTable";
 import Title from "@/components/common/Title";
 import AllMembersSearchInfoContextProvider from "@/contexts/AllMembersSearchInfoContext";
 
@@ -8,8 +8,8 @@ export default function AllMembersPage() {
     <>
       <Title variant={"allMember"} descriptionText={"전체 커뮤니티 멤버 정보를 관리합니다."} />
       <AllMembersSearchInfoContextProvider>
-        <Header variant={"allMember"} />
-        <AllMemberInfoTable />
+        <AllMembersHeader />
+        <AllMembersInfoTable />
       </AllMembersSearchInfoContextProvider>
     </>
   );
