+++ src/pages/PendingMembersPage.tsx
@@ -1,5 +1,5 @@
-import Header from "@/components/common/Header";
-import PendingMemberInfoTable from "@/components/common/Table/PendingMemberInfoTable";
+import PendingMembersHeader from "@/components/common/Header/PendingMembersHeader";
+import PendingMembersInfoTable from "@/components/common/Table/PendingMembersInfoTable";
 import Title from "@/components/common/Title";
 import PendingMembersSearchInfoContextProvider from "@/contexts/PendingMembersSearchInfoContext";
 
@@ -11,8 +11,8 @@ export default function PendingMembersPage() {
         descriptionText={"소셜 로그인을 완료한 멤버를 관리합니다. "}
       />
       <PendingMembersSearchInfoContextProvider>
-        <Header variant={"pendingMember"} />
-        <PendingMemberInfoTable />
+        <PendingMembersHeader />
+        <PendingMembersInfoTable />
       </PendingMembersSearchInfoContextProvider>
     </>
   );
